Separate generator -> Duplex utility
