+++ lib/stdio/async.js
@@ -4,7 +4,7 @@ import {Readable, Writable} from 'node:stream';
 import mergeStreams from '@sindresorhus/merge-streams';
 import {handleInput} from './handle.js';
 import {TYPE_TO_MESSAGE} from './type.js';
-import {generatorToTransformStream, pipeGenerator} from './generator.js';
+import {generatorToDuplexStream, pipeGenerator} from './generator.js';
 
 // Handle `input`, `inputFile`, `stdin`, `stdout` and `stderr` options, before spawning, in async mode
 export const handleInputAsync = options => handleInput(addPropertiesAsync, options, false);
@@ -15,7 +15,7 @@ const forbiddenIfAsync = ({type, optionName}) => {
 
 const addPropertiesAsync = {
 	input: {
-		generator: generatorToTransformStream,
+		generator: generatorToDuplexStream,
 		fileUrl: ({value}) => ({value: createReadStream(value)}),
 		filePath: ({value}) => ({value: createReadStream(value.file)}),
 		webStream: ({value}) => ({value: Readable.fromWeb(value)}),
@@ -24,7 +24,7 @@ const addPropertiesAsync = {
 		uint8Array: ({value}) => ({value: Readable.from(Buffer.from(value))}),
 	},
 	output: {
-		generator: generatorToTransformStream,
+		generator: generatorToDuplexStream,
 		fileUrl: ({value}) => ({value: createWriteStream(value)}),
 		filePath: ({value}) => ({value: createWriteStream(value.file)}),
 		webStream: ({value}) => ({value: Writable.fromWeb(value)}),
